Se termino todos los Servicios

+++ src/main/java/ventaproducto/ventasproductos/dto/ItemPedido/ItemPedidoDto.java
@@ -1,13 +1,11 @@
 package ventaproducto.ventasproductos.dto.ItemPedido;
 
-import java.math.BigDecimal;
-
 import ventaproducto.ventasproductos.dto.Producto.ProductoDto;
 
 public record ItemPedidoDto(
             long id,
             ProductoDto productoDto,
             int cantidad,
-            BigDecimal precioUnitario) {
+            int precioUnitario) {
 
 }
