+++ src/main/java/ventaproducto/ventasproductos/dto/Producto/ProductoDto.java
@@ -1,11 +1,9 @@
 package ventaproducto.ventasproductos.dto.Producto;
 
-import java.math.BigDecimal;
-
 public record ProductoDto(
             Long id,
             String nombre,
-            BigDecimal price,
+            int price,
             int stock){
