+++ src/main/java/ventaproducto/ventasproductos/dto/Producto/ProductoDtoSave.java
@@ -1,11 +1,9 @@
 package ventaproducto.ventasproductos.dto.Producto;
 
-import java.math.BigDecimal;
-
 public record ProductoDtoSave (
         Long id,
         String nombre,
-        BigDecimal price,
+        int price,
         int stock){
     
 }
