+++ src/main/java/ventaproducto/ventasproductos/entities/ItemPedido.java
@@ -1,6 +1,5 @@
 package ventaproducto.ventasproductos.entities;
 
-import java.math.BigDecimal;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -43,5 +42,5 @@ public class ItemPedido {
     private int cantidad;
 
     @Column
-    private BigDecimal precioUnitario;
+    private int precioUnitario;
 }
